refactor/meta_voting: split meta_vote.rs into multiple submodules
