+++ src/meta_voting/mod.rs
@@ -1,0 +1,21 @@
+// Copyright 2018 MaidSafe.net limited.
+//
+// This SAFE Network Software is licensed to you under The General Public License (GPL), version 3.
+// Unless required by applicable law or agreed to in writing, the SAFE Network Software distributed
+// under the GPL Licence is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied. Please review the Licences for the specific language governing
+// permissions and limitations relating to use of the SAFE Network Software.
+
+mod bool_set;
+mod counts;
+mod elections;
+mod vote;
+
+pub use self::bool_set::BoolSet;
+pub use self::elections::MetaElections;
+pub use self::vote::{MetaVote, Step};
+
+use hash::Hash;
+use std::collections::BTreeMap;
+
+pub(super) type MetaVotes<P> = BTreeMap<Hash, BTreeMap<P, Vec<MetaVote>>>;
